+++ bin/release.sh
@@ -29,7 +29,7 @@ cd /app/bugsnag-js-performance
 npm ci
 
 # check if CDN packages changed – if they didn't we don't need to upload to the CDN
-BROWSER_PACKAGE_CHANGED=$(npx lerna changed --parseable | grep -c packages/platforms/js$ || test $? = 1;)
+BROWSER_PACKAGE_CHANGED=$(npx lerna changed --parseable | grep -c packages/platforms/browser$ || test $? = 1;)
 
 # increment package version numbers
 if [ -z "${RETRY_PUBLISH:-}" ]; then