+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 ### Removed
 
 - (browser) Remove the `networkInstrumentationIgnoreUrls` config option [#261](https://github.com/bugsnag/bugsnag-js-performance/pull/261)
+- Removed `@bugsnag/js-performance` meta package [#264](https://github.com/bugsnag/bugsnag-js-performance/pull/264)
 
 ## v0.2.0 (2023-07-10)