+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 - (browser) Calculate time origin on create as well as visibility change [#268](https://github.com/bugsnag/bugsnag-js-performance/pull/268)
 
+### Removed
+
+- Removed `@bugsnag/js-performance` meta package [#264](https://github.com/bugsnag/bugsnag-js-performance/pull/264)
+
 ## v0.3.0 (2023-07-17)
 
 ### Fixed