+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Fixed
 
 - (browser) Calculate time origin on create as well as visibility change [#268](https://github.com/bugsnag/bugsnag-js-performance/pull/268)
+- (browser) Only track network requests over http(s) [#275](https://github.com/bugsnag/bugsnag-js-performance/pull/275)
 
 ### Removed